move eliminated components to example folder
