+++ src/example-components/dynamicBackground/types/index.ts
